Cloud Functions/github-issue-proxy/requirements.txt
functions-framework==3.*
google-cloud-logging
requests==2.*
flask